Fix sign warnings for rexp.
rexp can be negative, so we need to cast `uint32_t precision`
to int32_t before performing arithmetic.

Later, after testing `rexp < 0`, we have
`const int32_t requiredFives = -rexp;`. Therefore, `requiredFives > 0`,
so we could cast and store uint32_t to match multipleOfPowerOf5()'s
second parameter. However, that would be inconsistent with
`int32_t requiredTwos`, so instead I'm introducing a value-preserving
cast when calling multipleOfPowerOf5().

+++ ryu/d2fixed.c
@@ -712,12 +712,11 @@ int d2exp_buffered_n(double d, uint32_t precision, char* result) {
     // Is m * 2^e2 * 10^(precision + 1 - exp) integer?
     bool trailingZeros;
     // precision was already increased by 1, so we don't need to write + 1 here.
-    const int32_t rexp = precision - exp;
+    const int32_t rexp = (int32_t) precision - exp;
     const int32_t requiredTwos = -e2 - rexp;
     if (rexp < 0) {
       const int32_t requiredFives = -rexp;
-      trailingZeros =
-          (requiredTwos <= 0 || (requiredTwos < 60 && multipleOfPowerOf2(m2, requiredTwos))) && multipleOfPowerOf5(m2, requiredFives);
+      trailingZeros = (requiredTwos <= 0 || (requiredTwos < 60 && multipleOfPowerOf2(m2, requiredTwos))) && multipleOfPowerOf5(m2, (uint32_t) requiredFives);
     } else {
       trailingZeros = requiredTwos <= 0 || (requiredTwos < 60 && multipleOfPowerOf2(m2, requiredTwos));
     }
